Create initial program structure
This commit adds a scaffold, including test runner and linter, for a simple program to parse a log file of a specified format into some aggregated statistics derived from the log data. It includes initial models for the core components: a Parser to parse the log file, line by line, extracting the relevant data into an in-memory hash; a Formatter, which will eventually print the relevant statistics to the console; and a Stats module to crunch the numbers. It also includes a simple LogParser service to act as an entry point for running the program.